common: move peer's verbs mocks to common
