+++ tests/unit/common/mocks-ibverbs.h
@@ -23,8 +23,19 @@ extern struct ibv_mr Ibv_mr;
 #define MOCK_VERBS		(struct ibv_context *)0x4E4B
 #define MOCK_COMP_CHANNEL	(struct ibv_comp_channel *)&Ibv_comp_channel
 #define MOCK_IBV_CQ		(struct ibv_cq *)&Ibv_cq
+#define MOCK_IBV_PD		(struct ibv_pd *)0x00D0
 #define MOCK_QP			(struct ibv_qp *)&Ibv_qp
 #define MOCK_MR			(struct ibv_mr *)&Ibv_mr
+
+struct ibv_alloc_pd_mock_args {
+	int validate_params;
+	struct ibv_pd *pd;
+};
+
+struct ibv_dealloc_pd_mock_args {
+	int validate_params;
+	int ret;
+};
 
 struct ibv_post_send_mock_args {
 	struct ibv_qp *qp;
